+++ src/ralph/__init__.py
@@ -1,3 +1,3 @@
 """Ralph module"""
 
-__version__ = "0.0.1"
+__version__ = "1.1.0"